Remove dead code related to hazptr_obj_reclaim.
In particular, remove the version of "flush" that takes
a specific reclaim function and compares things to it.
We can't easily replicate this functionality in the
"templated deleter" model, because the identity of the
deleter no longer appears anywhere in the linked list
managed by the hazptr_domain: all we have at that point
is a function pointer with signature `void(*)(hazptr_obj*)`
whose pointer-value depends on the type of the thing
being reclaimed, but does not depend on the value of
that thing's deleter.

I'm hoping that this functionality turns out not to be
super important. Certainly there don't seem to be any
tests for it at the moment (as removing it broke nothing).